End of section 17

+++ API/SignalR/MessageHub.cs
@@ -21,18 +21,22 @@ namespace API.SignalR
 
             var groupName = GetGroupName(Context.User.GetUsername(), otherUser);
             await Groups.AddToGroupAsync(Context.ConnectionId, groupName);
-            await AddToGroup(groupName);
+            var group = await AddToGroup(groupName);
+
+            await Clients.Group(groupName).SendAsync("UpdatedGroup", group);
 
             var messages = await messageRepository.GetMessageThread(Context.User.GetUsername()
                 , otherUser!);
             
-            await Clients.Group(groupName).SendAsync("ReceiveMessageThread", messages);
+            //await Clients.Group(groupName).SendAsync("ReceiveMessageThread", messages);
+            await Clients.Caller.SendAsync("ReceiveMessageThread", messages);
 
         }
 
         public override async Task OnDisconnectedAsync(Exception? exception)
         {
-            await RemoveFromMessageGroup();
+            var group = await RemoveFromMessageGroup();
+            await Clients.Group(group.Name).SendAsync("UpdatedGroup", group);
             await base.OnDisconnectedAsync(exception);
         }
 
@@ -71,7 +75,8 @@ namespace API.SignalR
                 if(connections != null && connections?.Count != null)
                 {
                     await presenceHub.Clients.Clients(connections).SendAsync("NewMessageReceived",
-                        new {username = sender.UserName, KnownAs = sender.KnownAs});
+                        new {username = sender.UserName, knownAs = sender.KnownAs});
+                
                 }
             }
 
@@ -83,7 +88,7 @@ namespace API.SignalR
             }
         }
 
-        private async Task<bool> AddToGroup(string groupName)
+        private async Task<Group> AddToGroup(string groupName)
         {
             var username = Context.User?.GetUsername() ?? throw new Exception("Cannot get username");
             var group = await messageRepository.GetMessageGroup(groupName);
@@ -97,17 +102,21 @@ namespace API.SignalR
 
             group.Connections.Add(connection);
 
-            return await messageRepository.SaveAllAsync();  
+            if (await messageRepository.SaveAllAsync()) return group;
+            
+            throw new HubException("Failed to join group");
         }
 
-        private async Task RemoveFromMessageGroup()
+        private async Task<Group> RemoveFromMessageGroup()
         {
+            var group = await messageRepository.GetGroupForConnection(Context.ConnectionId);
             var connection = await messageRepository.GetConnection(Context.ConnectionId);
-            if(connection != null)
+            if(connection != null && group != null)
             {
                 messageRepository.RemoveConnection(connection);
-                await messageRepository.SaveAllAsync();
+                if(await messageRepository.SaveAllAsync()) return group;
             }
+            throw new HubException("Failed to remove from group");
         }
 
         private string GetGroupName(string caller, string? other)
